fix: move NextTrack to sessionStateUseCase
